requirements.txt
# 文档转换调度系统依赖包
# 复刻MediaConvert的依赖结构，支持企业级功能

# ===== Web框架 =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===== 数据库支持 =====
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
aiomysql==0.2.0

# ===== 云存储支持 =====
boto3==1.34.0
botocore==1.34.0

# ===== HTTP客户端 =====
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# ===== 数据验证和序列化 =====
pydantic>=2.5.0,<3.0.0
pydantic-settings==2.1.0

# ===== 日志记录 =====
loguru==0.7.2

# ===== 工具库 =====
python-dotenv==1.0.0
click==8.1.7

# ===== 异步支持 =====
asyncio==3.4.3

# ===== 系统工具 =====
psutil==5.9.6

# ===== 文件格式支持 =====
python-magic==0.4.27

# ===== 安全 =====
cryptography>=41.0.0

# ===== PDF处理 - 基础镜像已包含MinerU =====
# mineru[core]==2.1.9  # 基础镜像已包含

# ===== 开发和测试（可选） =====
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0